refactor(tabs/library): move state init to model
